+++ game/plugins/src/main/kotlin/gg/rsmod/plugins/content/mechanics/trading/impl/TradeSession.kt
@@ -289,4 +289,65 @@ class TradeSession(private val player: Player, private val partner: Player) {
         player.closeInterface(OVERLAY_INTERFACE)
     }
 
+    companion object {
+        /**
+         * The inventory overlay interface
+         */
+        const val OVERLAY_INTERFACE = 336
+
+        /**
+         * The primary trade screen interface
+         */
+        const val TRADE_INTERFACE = 335
+
+        /**
+         * The child id of this player's trade offer
+         */
+        const val PLAYER_TRADE_CHILD = 25
+
+        /**
+         * The child id of the partner's trade offer
+         */
+        private const val PARTNER_TRADE_CHILD = 28
+
+        /**
+         * The hash of this player's trade offer component
+         */
+        val PLAYER_TRADE_HASH = TRADE_INTERFACE.getInterfaceHash(PLAYER_TRADE_CHILD)
+
+        /**
+         * The hash of the partner's trade offer component
+         */
+        val PARTNER_TRADE_HASH = TRADE_INTERFACE.getInterfaceHash(PARTNER_TRADE_CHILD)
+
+        /**
+         * The progress trade interface
+         */
+        const val ACCEPT_INTERFACE = 334
+
+        /**
+         * The message that is shown to the partner when a player declined a trade
+         */
+        const val TRADE_DECLINED_MESSAGE = "Other player declined trade."
+
+        /**
+         * The container key for the trade accept screen
+         */
+        const val ACCEPT_CONTAINER_KEY = 90
+
+        /**
+         * The container key for the player's inventory overlay
+         */
+        const val PLAYER_INVENTORY_KEY = 93
+
+        /**
+         * The container key for this player's trade offer
+         */
+        const val PLAYER_CONTAINER_KEY = 24
+
+        /**
+         * The container key for the partner's trade offer
+         */
+        const val PARTNER_CONTAINER_KEY = 23
+    }
 }
